[OPENENGSB-3446] fixed pullrequest bugs in EDB JPA
