+++ components/edb/src/main/java/org/openengsb/core/edb/jpa/internal/JPACommit.java
@@ -43,7 +43,6 @@ import org.slf4j.LoggerFactory;
 @Entity
 @Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
 public class JPACommit extends VersionedEntity implements EDBCommit {
-
     private static final Logger LOGGER = LoggerFactory.getLogger(JPACommit.class);
 
     @Column(name = "COMMITER", length = 50)
