+++ components/edb/src/main/java/org/openengsb/core/edb/jpa/internal/JPAObject.java
@@ -38,7 +38,6 @@ import org.apache.openjpa.persistence.jdbc.Index;
  * class.
  */
 public class JPAObject extends VersionedEntity {
-
     @OneToMany(cascade = CascadeType.ALL, fetch = FetchType.EAGER, mappedBy = "owner")
     private List<JPAEntry> entries;
     @Column(name = "TIME")
