+++ components/edb/src/main/java/org/openengsb/core/edb/jpa/internal/dao/JPADao.java
@@ -38,6 +38,9 @@ public interface JPADao {
      */
     JPAHead getJPAHead(long timestamp) throws EDBException;
 
+    /**
+     * Loads the JPAHead with the given timestamp from the given stage.
+     */
     JPAHead getJPAHead(long timestamp, String sid) throws EDBException;
 
     /**
@@ -45,6 +48,9 @@ public interface JPADao {
      */
     List<JPAObject> getJPAObjectHistory(String oid) throws EDBException;
 
+    /**
+     * Returns the history (all objects) of a given object in the given stage.
+     */
     List<JPAObject> getJPAObjectHistory(String oid, String sid) throws EDBException;
 
     /**
@@ -52,6 +58,9 @@ public interface JPADao {
      */
     List<JPAObject> getJPAObjectHistory(String oid, long from, long to) throws EDBException;
 
+    /**
+     * Returns the history (between from and to) of a given object from the given stage.
+     */
     List<JPAObject> getJPAObjectHistory(String oid, String sid, long from, long to) throws EDBException;
 
     /**
@@ -59,6 +68,9 @@ public interface JPADao {
      */
     JPAObject getJPAObject(String oid, long timestamp) throws EDBException;
 
+    /**
+     * Returns a staged JPAObject with the given timestamp
+     */
     JPAObject getJPAObject(String oid, String sid, long timestamp) throws EDBException;
     
     /**
@@ -66,6 +78,9 @@ public interface JPADao {
      */
     JPAObject getJPAObject(String oid) throws EDBException;
 
+    /**
+     * Returns the newest staged JPAObject with the given oid
+     */
     JPAObject getJPAObject(String oid, String sid) throws EDBException;
     
     /**
@@ -73,6 +88,9 @@ public interface JPADao {
      */
     List<JPAObject> getJPAObjects(List<String> oids) throws EDBException;
 
+    /**
+     * Returns the newest staged JPAObjects with the given oids
+     */
     List<JPAObject> getJPAObjects(List<String> oids, String sid) throws EDBException;
 
     /**
@@ -80,6 +98,10 @@ public interface JPADao {
      */
     List<JPACommit> getJPACommit(String oid, long from, long to) throws EDBException;
 
+    /**
+     * Returns all commits which are involved with the given oid which are between from and to 
+     * and in the given stage
+     */
     List<JPACommit> getJPACommit(String oid, String sid, long from, long to) throws EDBException;
     
     /**
@@ -87,7 +109,11 @@ public interface JPADao {
      * this revision
      */
     JPACommit getJPACommit(String revision) throws EDBException;
-    
+
+    /**
+     * Returns the commit object for the given revision string in the given stage. 
+     * Throws an EDBException in case of no commit present for this revision
+     */
     JPACommit getJPACommit(String revision, String sid) throws EDBException;
 
     /**
@@ -95,6 +121,9 @@ public interface JPADao {
      */
     List<String> getResurrectedOIDs() throws EDBException;
 
+    /**
+     * Returns a list of oids from the staged JPAObjects which has been resurrected
+     */
     List<String> getResurrectedOIDs(String sid) throws EDBException;
 
     /**
@@ -102,6 +131,9 @@ public interface JPADao {
      */
     List<JPACommit> getJPACommit(long timestamp) throws EDBException;
 
+    /**
+     * Loads a JPACommit with the given timestamp from the given stage
+     */
     List<JPACommit> getJPACommit(long timestamp, String sid) throws EDBException;
 
     /**
@@ -109,7 +141,11 @@ public interface JPADao {
      * object.
      */
     List<CommitMetaInfo> getRevisionsOfMatchingCommits(CommitQueryRequest request) throws EDBException;
-    
+
+    /**
+     * Returns a list of commit meta information of all commits which are matching the request of the given request
+     * object and in the given stage.
+     */
     List<CommitMetaInfo> getRevisionsOfMatchingCommits(CommitQueryRequest request, String sid) throws EDBException;
 
     /**
@@ -117,6 +153,10 @@ public interface JPADao {
      */
     List<JPACommit> getCommits(Map<String, Object> param) throws EDBException;
 
+    /**
+     * Get all commits which are given with the param map and in the given stage. 
+     * In the map there are values like commiter, role, etc.
+     */
     List<JPACommit> getCommits(Map<String, Object> param, String sid) throws EDBException;
 
     /**
@@ -124,13 +164,19 @@ public interface JPADao {
      */
     JPACommit getLastCommit(Map<String, Object> param) throws EDBException;
 
+    /**
+     * like getCommits, but it returns only the newest commit from the stage
+     */
     JPACommit getLastCommit(Map<String, Object> param, String sid) throws EDBException;
     
     /**
      * Returns a list of JPAObjects which match to the parameters in the given query request
      */
     List<JPAObject> query(QueryRequest request) throws EDBException;
-    
+
+    /**
+     * Returns a list of staged JPAObjects which match to the parameters in the given query request
+     */
     List<JPAObject> query(QueryRequest request, String sid) throws EDBException;
 
     /**
@@ -138,5 +184,9 @@ public interface JPADao {
      */
     Integer getVersionOfOid(String oid) throws EDBException;
 
+    /**
+     * Returns the version of the element under the given oid and from the given stage. 
+     * If oid isn't existing, 0 is returned.
+     */
     Integer getVersionOfOid(String oid, String sid) throws EDBException;
 }
