+++ components/edb/src/main/resources/META-INF/persistence.xml
@@ -28,7 +28,7 @@
     <class>org.openengsb.core.edb.jpa.internal.JPACommit</class>
     <class>org.openengsb.core.edb.jpa.internal.JPAObject</class>
     <class>org.openengsb.core.edb.jpa.internal.JPAEntry</class>
-	<class>org.openengsb.core.edb.jpa.internal.JPAStage</class>
+    <class>org.openengsb.core.edb.jpa.internal.JPAStage</class>
     <exclude-unlisted-classes>true</exclude-unlisted-classes>
 
     <properties>
