+++ components/edb/src/test/java/org/openengsb/core/edb/jpa/internal/EDBStageFunctionTest.java
@@ -20,8 +20,8 @@ package org.openengsb.core.edb.jpa.internal;
 public class EDBStageFunctionTest extends AbstractEDBFunctionTest {
 
     @Override
-    protected Tools initTools() {
-        return new StageTestTools();
+    protected EDBTestDataGenerator initTools(TestEDBService db) {
+        return new EDBStagedTestData(db);
     }
 
 }
